requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
openai>=1.12.0
pydantic>=2.6.0
pytest>=8.0.0
python-jose>=3.3.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
boto3>=1.34.0
pandas>=2.2.0
numpy>=1.26.0
pypdf>=4.0.0
pdfplumber>=0.10.3
tiktoken>=0.6.0
psycopg2-binary>=2.9.9 